+++ packages/utils/src/index.ts
@@ -21,6 +21,7 @@ export * from "./frontmatter";
 export * from "./fs-error";
 export * from "./fs-optional";
 export * from "./glob";
+export * from "./github-check-run";
 export * from "./json";
 export * from "./json-parse";
 export * from "./jsonl-bytes";